API - refactor createGameKey to associate with phaser.create
