+++ src/components/HomePage.js
@@ -50,16 +50,6 @@ function HomePage(props) {
     setJoinGameSettings(true);
   };
 
-  const generateGameCode = () => {
-    let result = "";
-    let characters =
-      "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
-    let charactersLength = characters.length;
-    for (let i = 0; i < 5; i++) {
-      result += characters.charAt(Math.floor(Math.random() * charactersLength));
-    }
-    setGameCode(result);
-  };
   return (
     <div className={classes.gameHero}>
       <div className={classes.infoSection}>
